# perf: make room notification caching up to 4000% faster