The room notification modes will only change when the push rules have changed; otherwise, recomputing them is a no-op, and only wastes time.

Here's a distribution of the worst processing times under this method, as measured in `multiverse` with some big test account running (fresh, i.e. after an initial response) for a few minutes:

Before:
0.036192141s
0.037368775s
0.039177686s
0.039883797s
0.041679405s

After:
0.000000249s
0.000000397s
0.000000899s
0.001080307s
0.001340611s

The new distribution makes sense: the push rules event comes only once, in the initial response, so they're processed only once. Then, there's no need to process it ever again (in my session I've never changed the processing times), so it's almost instantaneous the subsequent times.

Found thanks to https://github.com/matrix-org/matrix-rust-sdk/pull/5612